Gameの役割じゃなくEngineの役割だと思う処理を移動
